backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
google-cloud-speech==2.23.0
google-cloud-aiplatform>=1.71.0
vertexai>=1.71.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
